// src/apis/generated/model/chatMessageRes.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * 동행 API
 * 동행 서비스 API 입니다.
 * OpenAPI spec version: v1
 */
import type { ChatMessageResRole } from "./chatMessageResRole";

export interface ChatMessageRes {
  role?: ChatMessageResRole;
  content?: string;
  createdAt?: string;
}