// src/apis/generated/model/chatMessageResRole.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * 동행 API
 * 동행 서비스 API 입니다.
 * OpenAPI spec version: v1
 */

export type ChatMessageResRole =
  (typeof ChatMessageResRole)[keyof typeof ChatMessageResRole];

export const ChatMessageResRole = {
  USER: "USER",
  AI: "AI",
} as const;